refactor: update layout to two panel steps
