+++ wordpress-scripts-webpack.config.js
@@ -51,7 +51,7 @@ module.exports = {
         classicFormDesignCss: srcPath('DonationForms/FormDesigns/ClassicFormDesign/css/main.scss'),
         classicFormDesignJs: srcPath('DonationForms/FormDesigns/ClassicFormDesign/js/main.ts'),
         multiStepFormDesignCss: srcPath('DonationForms/FormDesigns/MultiStepFormDesign/css/main.scss'),
-        twoColumnFormDesignCss: srcPath('DonationForms/FormDesigns/TwoColumnFormDesign/css/main.scss'),
+        twoPanelStepsFormLayoutCss: srcPath('DonationForms/FormDesigns/twoPanelStepsFormLayout/css/main.scss'),
         donationConfirmationReceiptApp: srcPath('DonationForms/resources/receipt/DonationConfirmationReceiptApp.tsx'),
         baseFormDesignCss: srcPath('DonationForms/resources/styles/base.scss'),
         formBuilderApp: srcPath('FormBuilder/resources/js/form-builder/src/index.tsx'),
